improves program integration, separates gem file and creates beta test file

+++ twitter_connection.rb
@@ -1,9 +1,6 @@
-require 'twitter'
-require 'yaml'
 
 class TwitterConnection
-  attr_reader :account, :client, :since
-  # include Singleton
+  attr_reader :account, :client
 
   def initialize(account = 'realDonaldTrump')
     @account = account
@@ -14,7 +11,16 @@ class TwitterConnection
   def tweets_to_doc(num = 200)
     tweets_raw = @client.user_timeline(@account, count: num)
     tweets = convert_tweets(tweets_raw)
-    File.write("#{@account}.yml", YAML.dump(tweets))
+    tweets = tweets.sort_by &:time
+
+    if File.exist?("./#{@account}.yml")
+      # since = YAML.load_file("#{@account}.yml").last.time
+      # tweets = tweets.select {|tweet| tweet.time > since}
+      # File.write("#{@account}.yml", YAML.dump(tweets), IO::SEEK_SET)
+    else
+      File.write("#{@account}.yml", YAML.dump(tweets))
+    end
+
   end
 
 
@@ -26,10 +32,10 @@ class TwitterConnection
 
   def client_access
     Twitter::REST::Client.new do |config|
-      config.consumer_key        = "[redacted]"
-      config.consumer_secret     = "[redacted]"
-      config.access_token        = "[redacted]"
-      config.access_token_secret = "[redacted]"
+      config.consumer_key        = "vJAwBD7ffyXMKtTe24LMPIA90"
+      config.consumer_secret     = "TKWrI5ctEZ4a2Amw4Ks7WNGMdsGsWfylkatt4y5k1g0UECenjs"
+      config.access_token        = "180542222-RyJ1oOMUsttp6epfKJAmMRJlrBwZm55orKFCfNqI"
+      config.access_token_secret = "90xcQjmVwLw5HORcXzJ7ZExOAlKSgQHQHnJZttSGMZUU7"
     end
   end
